fix: don't coerce u128 to u64

+++ state-chain/runtime/src/runtime_apis.rs
@@ -17,6 +17,6 @@ decl_runtime_apis!(
 		fn cf_authority_emission_per_block() -> u64;
 		fn cf_backup_emission_per_block() -> u64;
 		/// Returns the flip supply in the form [total_issuance, offchain_funds]
-		fn cf_flip_supply() -> (u64, u64);
+		fn cf_flip_supply() -> (u128, u128);
 	}
 );
